requirements.txt
blinker==1.9.0
click==8.1.7
colorama==0.4.6
flask==3.1.0
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==3.0.2
mysql-connector-python==8.0.33
protobuf==3.20.3
werkzeug==3.1.3
zipp==3.21.0
